refactor(dh): move routing artifacts in new library
