+++ libs/dh/metering-point/feature-overview/src/lib/breadcrumb/dh-breadcrumb.component.ts
@@ -27,8 +27,7 @@ import { TranslocoModule } from '@ngneat/transloco';
 import { WattIconModule } from '@energinet-datahub/watt';
 import { MeteringPointCimDto } from '@energinet-datahub/dh/shared/data-access-api';
 import { DhIsParentPipeScam } from '@energinet-datahub/dh/metering-point/shared/ui-util';
-
-import { dhMeteringPointPath } from '../routing/dh-metering-point-path';
+import { dhMeteringPointPath } from '@energinet-datahub/dh/metering-point/routing';
 
 export interface MeteringPointTranslationKeys {
   meteringMethod: string;
