+++ libs/dh/metering-point/feature-overview/src/lib/dh-metering-point-feature-overview.module.ts
@@ -18,11 +18,14 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
 import {
+  dhMeteringPointIdParam,
+  DhMeteringPointOverviewGuard,
+} from '@energinet-datahub/dh/metering-point/routing';
+
+import {
   DhMeteringPointOverviewComponent,
   DhMeteringPointOverviewScam,
 } from './dh-metering-point-overview.component';
-import { dhMeteringPointIdParam } from './routing/dh-metering-point-id-param';
-import { DhMeteringPointOverviewGuard } from './routing/dh-metering-point-overview.guard';
 
 const routes: Routes = [
   {
