+++ libs/dh/metering-point/feature-overview/src/lib/dh-metering-point-overview.component.spec.ts
@@ -27,10 +27,10 @@ import {
 } from '@ngworker/spectacular';
 import { getTranslocoTestingModule } from '@energinet-datahub/dh/shared/test-util-i18n';
 import { DhApiModule } from '@energinet-datahub/dh/shared/data-access-api';
+import { dhMeteringPointPath } from '@energinet-datahub/dh/metering-point/routing';
 
 import { DhMeteringPointFeatureOverviewModule } from './dh-metering-point-feature-overview.module';
 import { DhMeteringPointOverviewComponent } from './dh-metering-point-overview.component';
-import { dhMeteringPointPath } from './routing/dh-metering-point-path';
 
 describe(DhMeteringPointOverviewComponent.name, () => {
   beforeEach(async () => {
