+++ libs/dh/metering-point/feature-overview/src/lib/not-found/dh-metering-point-not-found.component.ts
@@ -18,13 +18,12 @@ import { ChangeDetectionStrategy, Component, NgModule } from '@angular/core';
 import { Router } from '@angular/router';
 import { TranslocoModule } from '@ngneat/transloco';
 
-import { dhMeteringPointSearchPath } from '@energinet-datahub/dh/metering-point/feature-search';
+import { dhMeteringPointSearchPath } from '@energinet-datahub/dh/metering-point/routing';
 import {
   WattButtonModule,
   WattEmptyStateModule,
 } from '@energinet-datahub/watt';
-
-import { dhMeteringPointPath } from '../routing/dh-metering-point-path';
+import { dhMeteringPointPath } from '@energinet-datahub/dh/metering-point/routing';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
