+++ libs/dh/metering-point/feature-search/src/index.ts
@@ -15,4 +15,3 @@
  * limitations under the License.
  */
 export * from './lib/dh-metering-point-search.component';
-export * from './lib/routing/dh-metering-point-search-path';
