+++ libs/dh/metering-point/routing/src/index.ts
@@ -14,3 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+export * from './lib/dh-metering-point-id-param';
+export * from './lib/dh-metering-point-path';
+export * from './lib/dh-metering-point-search-path';
+export * from './lib/dh-metering-point-overview.guard';
